Fix a race condition (#5)
When values are added and the topic is closed

  closed := make(chan struct{})
  top := topic.New(topic.WithClosed(closed))
  top.Publish("foo")
  close(closed)

It was pseudo random, if a receiver gets the value foo.

This fix makes sure, that a receiver always gets all values before the topic
is closed. It is still pseudo random, if it gets values that where added after the topic was closed.